finalized

+++ app/Http/Controllers/ListingController.php
@@ -70,7 +70,11 @@ class ListingController extends Controller
     //Update listing data
     public function update(Request $request, Listing $listing)
         
-    {   
+    {
+        //make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403,'Unauthorized Action');
+        }   
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -92,8 +96,17 @@ class ListingController extends Controller
 
     }
     public function destroy(Listing $listing)
-    {
+    {   
+        //make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403,'Unauthorized Action');
+        }
         $listing->delete();
         return redirect('/')->with('message','Listing deleted successfully');
+    }
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()
+    ->get()]);
     }
 }
